Refactor Bus and BusDevice
Bus has been refactored to store a Vec of BusDevices, and the BusDevice
enum now stores ArcMutexes of the actual devices. This does not change
any of the existing functionality, however it additionally allows the
devices to be shared.

This is needed for vCPU hotplugging, as the CpuContainer struct has to
be accessed by both the MMIODeviceManager and ACPIDeviceManager.

Signed-off-by: James Curtis <jxcurtis@amazon.co.uk>