Add snapshot save and restore functionality to CpuContainer
Implement Persist and the required functions for saving and restoring
snapshots

Note: Currently when saving a snapshot, custom CPU templates are not
saved. This is because, up until now, it has only been needed at
initialization. With the implementation of hotplugging however, CPU
templates and the vCPU config are now required for the lifetime of the
VMM. As a result of this, hotplugging will throw an error after
restoring from snapshot.

Signed-off-by: James Curtis <jxcurtis@amazon.co.uk>